+++ srcs/shift.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   shift.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jleroux <marvin@42lausanne.ch>             +#+  +:+       +#+        */
+/*   By: jleroux <jleroux@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/29 10:59:29 by jleroux           #+#    #+#             */
-/*   Updated: 2022/06/29 10:59:35 by jleroux          ###   ########.fr       */
+/*   Updated: 2022/06/29 13:45:31 by jleroux          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	shift_left(int *arr, int size)
+void	shift_left(long int *arr, int size)
 {
 	int	i;
 
@@ -22,7 +22,7 @@ void	shift_left(int *arr, int size)
 	}
 }
 
-void	shift_right(int *arr, int size)
+void	shift_right(long int *arr, int size)
 {
 	int	i;
